requirements.txt
flask
flask_session
gunicorn
pillow